backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# LangChain & LangGraph
langchain>=0.1.0
langchain-text-splitters>=1.0.0
langgraph>=0.0.40
langchain-openai>=0.0.5

# Vector Store
faiss-cpu>=1.7.4

# Cloud Storage
google-cloud-storage>=2.13.0

# Atlassian Integration
atlassian-python-api>=3.41.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.25.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-multipart>=0.0.6
